[Refactor] Code quality.
